+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Notifications and Requests:
 Types:
 - Renamed `TextDocumentPosition` to `TextDocumentPositionParams`.
 - Don't implement `TextDocumentIdentifier` on `FileEvent`.
+- `CompletionItem` -- Add an `additionalTextEdits` property.
 
 ## [1.0.1] - 2023-05-16